Create Delete duplicate elements of array.c
Write a C program to delete duplicate elements from an array using the following logic:

a. Take as input a number n: the total number of elements in the array and n number of integer elements into the array.
b. For every integer, count how many duplicates there are in the array.
c. Remove/Reduce duplicates in increasing order for every unique element, i.e., by 1 for the first element (which has duplicates), by 2 for the second, by 3 for the third, and so on .
d. If the number of duplicates is less than the required no. of deletions, delete all except the original element.
e. Display the array elements after removing the duplicates using the above logic.